* 'master' of github.com:Warzone2100/warzone2100:
  Remove use of %zd and %zu, since Windows felt like using %Id and %Iu instead.
  Try to fix isatty linkage errors on Windows.

+++ src/effects.cpp
@@ -312,7 +312,7 @@ static EFFECT *Effect_malloc(void)
 		/* Allocate new effect chunk */
 		EffectChunk *chunk = (EffectChunk *)calloc(1, sizeof(EffectChunk));
 
-		debug(LOG_MEMORY, "%zd effects in use, allocating %d extra", activeList.num, EFFECT_CHUNK_SIZE);
+		debug(LOG_MEMORY, "%lu effects in use, allocating %d extra", (unsigned long)activeList.num, EFFECT_CHUNK_SIZE);
 
 		/* Deal with out-of-memory conditions */
 		if (chunk == NULL) {
